requirements.txt
django-filter==1.0.4
djangorestframework==3.6.3
